Added: workable TrackController and used some routes in App for test
In this commit, just made a workable TrackController that can
automatically include related Class (according to record_type) and its
record_id, in this way, this tracker is not only focused on specific
type but universal type if the Model has already defined.

And also, updated TrackerSchema to fit BrowserClient info.